Simulation flow update
