+++ scripts/lib.py
@@ -1,0 +1,148 @@
+"""
+Copyright 2019 Google LLC
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+Parse the regression testlist in YAML format
+"""
+
+import os
+import random
+import sys
+import subprocess
+import time
+import yaml
+
+def read_yaml(yaml_file):
+  """ Read YAML file to a dictionary
+
+  Args:
+    yaml_file : YAML file
+
+  Returns:
+    yaml_data : data read from YAML in dictionary format
+  """
+  with open(yaml_file, "r") as f:
+    try:
+      yaml_data = yaml.safe_load(f)
+    except yaml.YAMLError as exc:
+      print(exc)
+      sys.exit(1)
+  return yaml_data
+
+
+def get_env_var(var):
+  """Get the value of environment variable
+
+  Args:
+    var : Name of the environment variable
+
+  Returns:
+    val : Value of the environment variable
+  """
+  try:
+    val = os.environ[var]
+  except KeyError:
+    print ("Please set the environment variable %0s" % var)
+    sys.exit(1)
+  return val
+
+
+def get_seed(seed):
+  """Get the seed to run the generator
+
+  Args:
+    seed : input seed
+
+  Returns:
+    seed to run instruction generator
+  """
+  if seed >= 0:
+    return seed
+  else:
+    return random.getrandbits(32)
+
+
+def run_cmd(cmd, verbose = 0):
+  """Run a command and return output
+
+  Args:
+    cmd : shell command to run
+
+  Returns:
+    command output
+  """
+  try:
+    ps = subprocess.Popen(cmd,
+                          shell=True,
+                          universal_newlines=True,
+                          stdout=subprocess.PIPE,
+                          stderr=subprocess.STDOUT)
+  except subprocess.CalledProcessError as exc:
+    print(ps.communicate()[0])
+    sys.exit(1)
+  output = ps.communicate()[0]
+  if verbose:
+    print(output)
+  return output
+
+
+def run_parallel_cmd(cmd_list, verbose = 0):
+  """Run a list of commands in parallel
+
+  Args:
+    cmd_list: command list
+
+  Returns:
+    command output
+  """
+  children = []
+  for cmd in cmd_list:
+    ps = subprocess.Popen(cmd,
+                          shell=True,
+                          universal_newlines=True,
+                          stdout=subprocess.PIPE,
+                          stderr=subprocess.STDOUT)
+    children.append(ps)
+  for i in range(len(children)):
+    print("Command progress: %d/%d" % (i, len(children)))
+    if verbose:
+      print("Waiting for command: %s" % cmd_list[i])
+    output = children[i].communicate()[0]
+    if verbose:
+      print(output)
+  # Restore stty setting otherwise the terminal may go crazy
+  os.system("stty sane")
+
+
+def process_regression_list(testlist, test, iterations, matched_list):
+  """ Get the matched tests from the regression test list
+
+  Args:
+    testlist     : Regression test list
+    test         : Test to run, "all" means all tests in the list
+    iterations   : Number of iterations for each test
+
+  Returns:
+    matched_list : A list of matched tests
+  """
+  print("Processing regression test list : %s, test: %s" % (testlist, test))
+  yaml_data = read_yaml(testlist)
+  for entry in yaml_data:
+    if (entry['test'] == test) or (test == "all"):
+      if iterations > 0:
+        entry['iterations'] = iterations
+      if entry['iterations'] > 0:
+        print ("Found matched tests: %s, iterations:%0d" %
+               (entry['test'], entry['iterations']))
+        matched_list.append(entry)
